Changer ID (Int to Long)

+++ src/main/java/com/cgi/dao/CategoryDao.java
@@ -2,6 +2,6 @@ package com.cgi.dao;
 
 import com.cgi.model.Category;
 
-public interface CategoryDao extends GenericDao<Category, Integer> {
+public interface CategoryDao extends GenericDao<Category, Long> {
 
 }
