+++ src/main/java/com/cgi/dao/CommentDao.java
@@ -2,6 +2,6 @@ package com.cgi.dao;
 
 import com.cgi.model.Comment;
 
-public interface CommentDao extends GenericDao<Comment, Integer> {
+public interface CommentDao extends GenericDao<Comment, Long> {
 
 }
