+++ src/main/java/com/cgi/dao/IdeaDao.java
@@ -2,6 +2,6 @@ package com.cgi.dao;
 
 import com.cgi.model.Idea;
 
-public interface IdeaDao extends GenericDao<Idea, Integer> {
+public interface IdeaDao extends GenericDao<Idea, Long> {
 
 }
