+++ src/main/java/com/cgi/dao/LoginDao.java
@@ -2,6 +2,6 @@ package com.cgi.dao;
 
 import com.cgi.model.Login;
 
-public interface LoginDao extends GenericDao<Login, Integer> {
+public interface LoginDao extends GenericDao<Login, Long> {
 
 }
